requirement.txt
kafka-python
request
